requirements.txt
apache-airflow-providers-sftp
apache-airflow-providers-amazon[s3]
pyodbc
rarfile
pandas
